+++ src/main/scala/at/ac/ait/Model.scala
@@ -1,7 +1,6 @@
 package at.ac.ait
 
 case class RegularInput(
-    addressPrefix: String,
     address: String,
     txHash: Array[Byte],
     value: Long,
@@ -12,7 +11,6 @@ case class RegularInput(
 )
 
 case class RegularOutput(
-    addressPrefix: String,
     address: String,
     txHash: Array[Byte],
     value: Long,
@@ -23,7 +21,22 @@ case class RegularOutput(
     n: Int
 )
 
-case class NormalizedAddress(id: Int, address: String)
+case class AddressId(
+    address: String,
+    addressId: Int
+)
+
+case class AddressIdByAddress(
+    addressPrefix: String,
+    address: String,
+    addressId: Int
+)
+
+case class AddressByIdGroup(
+    addressIdGroup: Int,
+    addressId: Int,
+    address: String
+)
 
 case class InputIdSet(inputs: Seq[Int]) extends Iterable[Int] {
   override def iterator = inputs.iterator
@@ -45,14 +58,14 @@ case class TxSummary(
 
 case class TxIdTime(height: Int, txHash: Array[Byte], timestamp: Int)
 
-case class Currency(satoshi: Long, eur: Double, usd: Double)
+case class Currency(value: Long, eur: Float, usd: Float)
 
-case class AddressSummary(totalReceived: Long, totalSpent: Long)
+case class AddressSummary(totalReceived: Currency, totalSpent: Currency)
 
 case class ClusterSummary(
     noAddresses: Int,
-    totalReceived: Long,
-    totalSpent: Long
+    totalReceived: Currency,
+    totalSpent: Currency
 )
 
 // raw schema tables
@@ -78,7 +91,9 @@ case class Transaction(
     txIndex: Long
 )
 
-case class ExchangeRates(height: Int, eur: Double, usd: Double)
+case class BlockTransactions(height: Int, txs: Seq[TxSummary])
+
+case class ExchangeRatesRaw(date: String, eur: Float, usd: Float)
 
 case class Tag(
     address: String,
@@ -87,16 +102,24 @@ case class Tag(
     tagpackUri: String,
     currency: String,
     lastmod: Int,
-    category: String
+    category: String,
+    abuse: Option[String]
+)
+
+case class SummaryStatisticsRaw(
+    id: String,
+    timestamp: Int,
+    noBlocks: Int,
+    noTxs: Long
 )
 
 // transformed schema tables
 
-case class BlockTransactions(height: Int, txs: Seq[TxSummary])
+case class ExchangeRates(height: Int, eur: Float, usd: Float)
 
 case class AddressTransactions(
-    addressPrefix: String,
-    address: String,
+    addressIdGroup: Int,
+    addressId: Int,
     txHash: Array[Byte],
     value: Long,
     height: Int,
@@ -105,8 +128,7 @@ case class AddressTransactions(
 )
 
 case class BasicAddress(
-    addressPrefix: String,
-    address: String,
+    addressId: Int,
     noIncomingTxs: Int,
     noOutgoingTxs: Int,
     firstTx: TxIdTime,
@@ -118,6 +140,7 @@ case class BasicAddress(
 case class Address(
     addressPrefix: String,
     address: String,
+    addressId: Int,
     noIncomingTxs: Int,
     noOutgoingTxs: Int,
     firstTx: TxIdTime,
@@ -129,19 +152,22 @@ case class Address(
 )
 
 case class AddressTags(
+    addressId: Int,
     address: String,
     label: String,
     source: String,
     tagpackUri: String,
     lastmod: Int,
-    category: String
+    category: String,
+    abuse: String
 )
 
-case class AddressCluster(addressPrefix: String, address: String, cluster: Int)
+case class AddressCluster(addressIdGroup: Int, addressId: Int, cluster: Int)
 
 case class ClusterAddresses(
+    clusterGroup: Int,
     cluster: Int,
-    address: String,
+    addressId: Int,
     noIncomingTxs: Int,
     noOutgoingTxs: Int,
     firstTx: TxIdTime,
@@ -174,7 +200,7 @@ case class BasicCluster(
 
 case class BasicClusterAddresses(
     cluster: Int,
-    address: String,
+    addressId: Int,
     noIncomingTxs: Int,
     noOutgoingTxs: Int,
     firstTx: TxIdTime,
@@ -184,6 +210,7 @@ case class BasicClusterAddresses(
 )
 
 case class Cluster(
+    clusterGroup: Int,
     cluster: Int,
     noAddresses: Int,
     noIncomingTxs: Int,
@@ -197,20 +224,23 @@ case class Cluster(
 )
 
 case class ClusterTags(
+    clusterGroup: Int,
     cluster: Int,
+    addressId: Int,
     address: String,
     label: String,
     source: String,
     tagpackUri: String,
     lastmod: Int,
-    category: String
+    category: String,
+    abuse: String
 )
 
 case class AddressRelations(
-    srcAddressPrefix: String,
-    dstAddressPrefix: String,
-    srcAddress: String,
-    dstAddress: String,
+    srcAddressIdGroup: Int,
+    srcAddressId: Int,
+    dstAddressIdGroup: Int,
+    dstAddressId: Int,
     srcProperties: AddressSummary,
     dstProperties: AddressSummary,
     noTransactions: Int,
@@ -226,7 +256,9 @@ case class PlainClusterRelations(
 )
 
 case class ClusterRelations(
+    srcClusterGroup: Int,
     srcCluster: Int,
+    dstClusterGroup: Int,
     dstCluster: Int,
     srcProperties: ClusterSummary,
     dstProperties: ClusterSummary,
@@ -236,10 +268,11 @@ case class ClusterRelations(
 
 case class SummaryStatistics(
     timestamp: Int,
-    noBlocks: Long,
+    noBlocks: Int,
     noTransactions: Long,
     noAddresses: Long,
     noAddressRelations: Long,
     noClusters: Long,
-    noTags: Long
+    noTags: Long,
+    bucketSize: Int
 )